Support --config (-c)
The config file, if specified, should be a file containing command
line switches as they would be passed to the program.

When the config file is loaded, the app will shove all of the bits
into @ARGV, tack on the original @ARGV, and reprocess it.

You may overload some of the command line arguments on the command
line. For example:

  Config (always verbose and test Some::Module)

    --verbose -a Some::Module

  Override verbose on the command line:

    p5-distribution-smoke -c config --noverbose <Some::Dist>